refactor: AbstractAuthenticationInterceptor 추출 및 구현체 추가
