+++ src/main/java/nextstep/auth/authentication/session/SessionAuthenticationInterceptor.java
@@ -1,38 +1,29 @@
 package nextstep.auth.authentication.session;
 
+import nextstep.auth.authentication.AbstractAuthenticationInterceptor;
 import nextstep.auth.authentication.AuthenticationConverter;
-import nextstep.auth.authentication.AuthenticationToken;
 import nextstep.auth.authentication.ProviderManager;
 import nextstep.auth.context.Authentication;
 import nextstep.auth.context.SecurityContext;
-import org.springframework.web.servlet.HandlerInterceptor;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import javax.servlet.http.HttpSession;
 
+import java.io.IOException;
+
 import static nextstep.auth.context.SecurityContextHolder.SPRING_SECURITY_CONTEXT_KEY;
 
-public class SessionAuthenticationInterceptor implements HandlerInterceptor {
-    public static final String USERNAME_FIELD = "username";
-    public static final String PASSWORD_FIELD = "password";
-
-    private final AuthenticationConverter converter;
-    private final ProviderManager providerManager;
+public class SessionAuthenticationInterceptor extends AbstractAuthenticationInterceptor {
 
     public SessionAuthenticationInterceptor(AuthenticationConverter converter, ProviderManager providerManager) {
-        this.converter = converter;
-        this.providerManager = providerManager;
+        super(converter, providerManager);
     }
 
     @Override
-    public boolean preHandle(HttpServletRequest request, HttpServletResponse response, Object handler) {
-        AuthenticationToken token = converter.convert(request);
-        Authentication authentication = providerManager.authenticate(token);
-
+    protected void afterAuthentication(HttpServletRequest request, HttpServletResponse response, Authentication authentication) throws IOException {
         HttpSession httpSession = request.getSession();
         httpSession.setAttribute(SPRING_SECURITY_CONTEXT_KEY, new SecurityContext(authentication));
         response.setStatus(HttpServletResponse.SC_OK);
-        return false;
     }
 }
