+++ src/test/java/nextstep/auth/unit/authentication/session/SessionAuthenticationConverterTest.java
@@ -36,8 +36,8 @@ class SessionAuthenticationConverterTest {
     private MockHttpServletRequest createMockRequest() {
         MockHttpServletRequest request = new MockHttpServletRequest();
         Map<String, String> params = new HashMap<>();
-        params.put(SessionAuthenticationInterceptor.USERNAME_FIELD, EMAIL);
-        params.put(SessionAuthenticationInterceptor.PASSWORD_FIELD, PASSWORD);
+        params.put(SessionAuthenticationConverter.USERNAME_FIELD, EMAIL);
+        params.put(SessionAuthenticationConverter.PASSWORD_FIELD, PASSWORD);
         request.addParameters(params);
         return request;
     }
